requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Security & Authentication
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Validation
pydantic[email]==2.5.3

# AI & Facial Recognition
opencv-contrib-python==4.9.0.80
numpy==1.26.3

# Utilities
python-dotenv==1.0.1
